Acceptence tests as an exponent
