+++ server/src/main/java/com/example/server/businessLayer/Policies/Discount/AmountOfItemCondition.java
@@ -1,12 +1,13 @@
 package com.example.server.businessLayer.Policies.Discount;
 
+import com.example.server.businessLayer.Item;
 import com.example.server.businessLayer.ShoppingBasket;
 
 public class AmountOfItemCondition extends Condition{
     private double amountNeeded;
-    private Class itemNeeded;
+    private Item itemNeeded;
 
-    public AmountOfItemCondition(double amountNeeded, Class itemNeeded) {
+    public AmountOfItemCondition(double amountNeeded, Item itemNeeded) {
         this.amountNeeded = amountNeeded;
         this.itemNeeded = itemNeeded;
     }
